requirements.txt
numpy
pandas
requests
librosa
scikit-learn
joblib
supabase
python-dotenv
scipy
ffmpeg-python
pydub
tensorflow
tensorflow-lite
seaborn
matplotlib